+++ src/main/java/kitchenpos/table/ui/request/OrderTableRequest.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.ui.request;
+package kitchenpos.table.ui.request;
 
-import kitchenpos.order.domain.NumberOfGuests;
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableEmpty;
+import kitchenpos.table.domain.NumberOfGuests;
+import kitchenpos.table.domain.TableEmpty;
 
 public class OrderTableRequest {
 	private final int numberOfGuests;
