+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -19,7 +19,7 @@ import kitchenpos.menu.ui.response.MenuResponse;
 import kitchenpos.menu.ui.response.ProductResponse;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.ui.response.OrderResponse;
-import kitchenpos.order.ui.response.OrderTableResponse;
+import kitchenpos.table.ui.response.OrderTableResponse;
 
 @DisplayName("주문 관련 기능")
 class OrderAcceptanceTest extends AcceptanceTest {
